requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
pillow==10.3.0
sqlparse==0.4.4
